Make locallyCrawled increment only when crawl is a success

+++ internal/pkg/frontier/item.go
@@ -16,7 +16,7 @@ type Item struct {
 	Redirect       int
 	URL            *url.URL
 	ParentItem     *Item
-	LocallyCrawled int
+	LocallyCrawled uint64
 }
 
 // NewItem initialize an *Item
